+++ src/BundlerMinifierVsix/BundlerMinifierPackage.cs
@@ -13,8 +13,9 @@ namespace BundlerMinifierVsix
     [InstalledProductRegistration("#110", "#112", BundlerMinifier.Constants.VERSION, IconResourceID = 400)]
     [ProvideMenuResource("Menus.ctmenu", 1)]
     [Guid(PackageGuids.guidBundlerPackageString)]
+    [ProvideOptionPage(typeof(Options), "Web", Vsix.Name, 101, 102, true, new[] { "bundle", "minify" }, ProvidesLocalizedCategoryName = false)]
     [ProvideAutoLoad(UIContextGuids80.SolutionExists)]
-       public sealed class BundlerMinifierPackage : Package
+    public sealed class BundlerMinifierPackage : Package
     {
         public static DTE2 _dte;
         public static Dispatcher _dispatcher;
@@ -22,13 +23,16 @@ namespace BundlerMinifierVsix
         public static BundlerMinifierPackage _instance;
         SolutionEvents _solutionEvents;
 
+        public static Options Options { get; private set; }
+
         protected override void Initialize()
         {
             _instance = this;
             _dte = GetService(typeof(DTE)) as DTE2;
             _dispatcher = Dispatcher.CurrentDispatcher;
             Package = this;
-   
+            Options = (Options)GetDialogPage(typeof(Options));
+
             Logger.Initialize(this, Vsix.Name);
 
             Events2 events = _dte.Events as Events2;
@@ -42,8 +46,11 @@ namespace BundlerMinifierVsix
             UpdateAllFiles.Initialize(this);
             BundleOnBuild.Initialize(this);
             RemoveBundle.Initialize(this);
-            ProjectEventCommand.Initialize(this);
             ClearOutputFiles.Initialize(this);
+            ToggleProduceOutput.Initialize(this);
+            OpenSettings.Initialize(this);
+            ProjectEventCommand.Initialize(this);
+            ConvertToGulp.Initialize(this);
 
             base.Initialize();
         }