+++ src/BundlerMinifierVsix/Resources/Files/gulpfile.js
@@ -1,0 +1,70 @@
+"use strict";
+
+var gulp = require("gulp"),
+    concat = require("gulp-concat"),
+    cssmin = require("gulp-cssmin"),
+    htmlmin = require("gulp-htmlmin"),
+    uglify = require("gulp-uglify"),
+    merge = require("merge-stream"),
+    del = require("del"),
+    bundleconfig = require("./bundleconfig.json");
+
+gulp.task("min", ["min:js", "min:css", "min:html"]);
+
+gulp.task("min:js", function () {
+    var tasks = getBundles(".js").map(function (bundle) {
+        return gulp.src(bundle.inputFiles, { base: "." })
+            .pipe(concat(bundle.outputFileName))
+            .pipe(uglify())
+            .pipe(gulp.dest("."));
+    });
+    return merge(tasks);
+});
+
+gulp.task("min:css", function () {
+    var tasks = getBundles(".css").map(function (bundle) {
+        return gulp.src(bundle.inputFiles, { base: "." })
+            .pipe(concat(bundle.outputFileName))
+            .pipe(cssmin())
+            .pipe(gulp.dest("."));
+    });
+    return merge(tasks);
+});
+
+gulp.task("min:html", function () {
+    var tasks = getBundles(".html").map(function (bundle) {
+        return gulp.src(bundle.inputFiles, { base: "." })
+            .pipe(concat(bundle.outputFileName))
+            .pipe(htmlmin({ collapseWhitespace: true, minifyCSS: true, minifyJS: true }))
+            .pipe(gulp.dest("."));
+    });
+    return merge(tasks);
+});
+
+gulp.task("clean", function () {
+    var files = bundleconfig.map(function (bundle) {
+        return bundle.outputFileName
+    });
+
+    return del(files);
+});
+
+gulp.task("watch", function () {
+    getBundles(".js").forEach(function (bundle) {
+        gulp.watch(bundle.inputFiles, ["min:js"]);
+    });
+
+    getBundles(".css").forEach(function (bundle) {
+        gulp.watch(bundle.inputFiles, ["min:css"]);
+    });
+
+    getBundles(".html").forEach(function (bundle) {
+        gulp.watch(bundle.inputFiles, ["min:html"]);
+    });
+});
+
+function getBundles(extension) {
+    return bundleconfig.filter(function (bundle) {
+        return new RegExp(extension).test(bundle.outputFileName);
+    });
+}